create

+++ python-more_data_structures/0-square_matrix_simple.py
@@ -5,7 +5,7 @@ def square_matrix_simple(matrix=[]):
 
     new_matrix = []
     for row in matrix:
-        new_row = [float(x * x) for x in row]
+        new_row = [int(x * x) for x in row]
         new_matrix.append(new_row)
 
     return new_matrix
